Update: Imergination SVD and monsoons maps
I originally was using the rain-gauge-adjusted "Final" dataset from
IMERG. This produced artifacts at the continent edges as the rain
gauge data only covers a crudely stenciled out land regiion.

I've replaced the input data with the Late IMERG data and the
artifacts are gone. The result is a lot smoother now (and to my eye
looks more correct)